+++ bot/minimax.py
@@ -4,7 +4,7 @@ from bot.gamestate import GameState, UltimateGameState
 
 # type import
 from typing import Optional
-from bot.gamestate import NumericType
+from bot.gamestate import TicTacToeBoard, UltimateTicTacToeBoard, PlayerCharacter, NumericType
 
 def minimax(game_state: GameState | UltimateGameState,
             maximizing: bool = True, *,
@@ -72,3 +72,38 @@ def minimax(game_state: GameState | UltimateGameState,
 #         recur = minimax(neighbour, not maximizing, depth=depth + 1)
 #         score = optimise_func(score, recur)
 #     return score
+
+def find_move_normal(board: TicTacToeBoard, turn: PlayerCharacter, *,
+                     kill_signal: list[bool])\
+    -> Optional[GameState]:
+    state = GameState(board, turn)
+    best_score = -inf
+    best_state = None
+    for new_state in state.expand_state():
+        if kill_signal[0]: break
+
+        score = minimax(new_state, False)
+        if best_score < score:
+            best_state = new_state
+            best_score = score
+    return best_state
+
+def find_move_ultimate(main_board: UltimateTicTacToeBoard,
+                       subboards: list[TicTacToeBoard],
+                       turn: PlayerCharacter,
+                       board_to_play: tuple[int, int] = None, *,
+                       kill_signal: list[bool],
+                       max_depth: int = 1000)\
+    -> Optional[UltimateGameState]:
+    state = UltimateGameState(main_board, subboards, turn,
+                              board_to_play=board_to_play)
+    best_score = -inf
+    best_state = None
+    for new_state in state.expand_state():
+        if kill_signal[0]: break
+
+        score = minimax(new_state, False, max_depth=max_depth)
+        if best_score < score:
+            best_state = new_state
+            best_score = score
+    return best_state
